clean typescript.ts and add guard clauses to identify transaction
