+++ app/ts/backgroundServiceWorker.ts
@@ -1,7 +1,7 @@
 import './background/background.js'
 import { getHtmlFile } from './background/backgroundUtils.js'
 import { clearTabStates } from './background/settings.js'
-import { sleep } from './utils/typescript.js'
+import { sleep } from './utils/sleep.js'
 
 browser.action.setPopup({ popup: getHtmlFile('popup') })
