+++ app/ts/utils/typescript.ts
@@ -50,49 +50,6 @@ export function assertIsObject(maybe: unknown): asserts maybe is Object {
 	if (!isObject(maybe)) throw new Error(`Expected object but got ${ typeof maybe }`)
 }
 
-type StorageKey = 'activeSigningAddress'
-	| 'activeSimulationAddress'
-	| 'addressInfos'
-	| 'page'
-	| 'useSignersAddressAsActiveAddress'
-	| 'websiteAccess'
-	| 'activeChain'
-	| 'simulationMode'
-	| 'pendingAccessRequests'
-	| 'contacts'
-	| 'makeMeRich'
-	| 'addressbookTabId'
-	| 'transactionsPendingForUserConfirmation'
-	| 'ChainChangeConfirmationPromise'
-	| 'PersonalSignPromise'
-	| 'InterceptorAccessRequestPromise'
-	| 'simulationResults'
-	| 'signerName'
-	| 'currentTabId'
-	| `tabState_${ number }`
-	| 'isConnectedToNode'
-	| 'ethereumSubscriptions'
-	| 'useTabsInsteadOfPopup'
-
-export async function browserStorageLocalGet(keys: StorageKey | StorageKey[]) {
-	return await browser.storage.local.get(keys) as Promise<Partial<Record<StorageKey, JSONEncodeable>>>
-}
-
-type JSONEncodeable = string | number | boolean | { [x: string]: JSONEncodeable } | ReadonlyArray<JSONEncodeable>
-
-export async function browserStorageLocalSet(key: StorageKey, value: JSONEncodeable) {
-	return await browser.storage.local.set({ [key]: value })
-}
-export async function browserStorageLocalSetKeys(items: Partial<Record<StorageKey, JSONEncodeable>>) {
-	return await browser.storage.local.set({ ...items })
-}
-
-export async function browserStorageLocalSingleGetWithDefault(key: StorageKey, valueIfMissing: unknown) {
-	const value = await browser.storage.local.get(key) as Partial<Record<StorageKey, unknown>>
-	if (value[key] === undefined) return valueIfMissing
-	return value[key]
-}
-
-export function sleep(milliseconds: number) {
-    return new Promise(resolve => setTimeout(resolve, milliseconds))
+export function createGuard<T, U extends T>(check: (maybe: T) => U | undefined): (maybe: T) => maybe is U {
+    return (maybe: T): maybe is U => check(maybe) !== undefined
 }
